Renames sorting helpers file
