+++ lib/normalize-options.js
@@ -1,4 +1,4 @@
-const { sortStrings } = require('./utils/sortStrings');
+const { sortStrings } = require('./utils/sort');
 
 const varPattern = /var\((--[\w-]+)\)/;
