+++ test/utils.test.js
@@ -1,5 +1,5 @@
 const valuesHaveSameUnits = require('../lib/utils/valuesHaveSameUnits');
-const { sortStrings, sortObjectsByProperty } = require('../lib/utils/sortStrings');
+const { sortStrings, sortObjectsByProperty } = require('../lib/utils/sort');
 const parseAtruleParams = require('../lib/utils/parseAtruleParams');
 const compareStrings = require('../lib/utils/compareStrings');
 
@@ -88,8 +88,13 @@ describe('Compare strings', () => {
 
 describe('Sort Objects', () => {
   test('Test one', () => {
-    // eslint-disable-next-line max-len
-    expect(sortObjectsByProperty([{ breakpoint: '1000px' }, { breakpoint: '600px' }, { breakpoint: '400px' }], 'breakpoint'))
+    /* eslint-disable function-paren-newline */
+    expect(
+      sortObjectsByProperty(
+        [{ breakpoint: '1000px' }, { breakpoint: '600px' }, { breakpoint: '400px' }],
+        'breakpoint',
+      ),
+    )
       .toEqual([{ breakpoint: '400px' }, { breakpoint: '600px' }, { breakpoint: '1000px' }]);
   });
 });
